devbox-app/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0
psycopg[binary]==3.2.1
pytest==8.3.2
pytest-django==4.8.0
